+++ components/examples/ExampleComponent.js
@@ -2,7 +2,7 @@ import './ExampleComponent.css';
 
 export default ({ children, title }) => (
 	<div>
-		<h1>{title}</h1>
+		<h1 data-testid="title">{title}</h1>
 		<p>Example Component</p>
 
 		{children}